import resources using docker client only for the right environments
