+++ agenta-backend/agenta_backend/services/templates_manager.py
@@ -1,10 +1,18 @@
 import json
+import backoff
 from typing import Any, Dict, List
-
+import httpx
+import os
 from agenta_backend.config import settings
-from agenta_backend.services import container_manager, db_manager
+from agenta_backend.services import db_manager
 from agenta_backend.utils import redis_utils
+from httpx import ConnectError, TimeoutException
+from asyncio.exceptions import CancelledError
 
+if os.environ["FEATURE_FLAG"] in ["oss", "cloud"]:
+    from agenta_backend.services import container_manager
+
+from typing import Union
 
 async def update_and_sync_templates(cache: bool = True) -> None:
     """
@@ -75,7 +83,7 @@ async def retrieve_templates_from_dockerhub_cached(cache: bool) -> List[dict]:
             return json.loads(cached_data.decode("utf-8"))
 
     # If not cached, fetch data from Docker Hub and cache it in Redis
-    response = await container_manager.retrieve_templates_from_dockerhub(
+    response = await retrieve_templates_from_dockerhub(
         settings.docker_hub_url,
         settings.docker_hub_repo_owner,
         settings.docker_hub_repo_name,
@@ -107,7 +115,7 @@ async def retrieve_templates_info_from_s3(
             return json.loads(cached_data)
 
     # If not cached, fetch data from Docker Hub and cache it in Redis
-    response = await container_manager.get_templates_info_from_s3(
+    response = await get_templates_info_from_s3(
         "https://llm-app-json.s3.eu-central-1.amazonaws.com/llm_info.json"
     )
 
@@ -115,3 +123,60 @@ async def retrieve_templates_info_from_s3(
     r.set("temp_data", json.dumps(response), ex=900)
     print("Using network call...")
     return response
+
+
+@backoff.on_exception(backoff.expo, (ConnectError, CancelledError), max_tries=5)
+async def retrieve_templates_from_dockerhub(
+    url: str, repo_owner: str, repo_name: str
+) -> Union[List[dict], dict]:
+    """
+    Business logic to retrieve templates from DockerHub.
+
+    Args:
+        url (str): The URL endpoint for retrieving templates. Should contain placeholders `{}`
+            for the `repo_owner` and `repo_name` values to be inserted. For example:
+            `https://hub.docker.com/v2/repositories/{}/{}/tags`.
+        repo_owner (str): The owner or organization of the repository from which templates are to be retrieved.
+        repo_name (str): The name of the repository where the templates are located.
+
+    Returns:
+        tuple: A tuple containing two values.
+    """
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get(
+            f"{url.format(repo_owner, repo_name)}/tags", timeout=10
+        )
+        if response.status_code == 200:
+            response_data = response.json()
+            return response_data
+
+        response_data = response.json()
+        return response_data
+
+
+@backoff.on_exception(
+    backoff.expo, (ConnectError, TimeoutException, CancelledError), max_tries=5
+)
+
+
+async def get_templates_info_from_s3(url: str) -> Dict[str, Dict[str, Any]]:
+    """
+    Business logic to retrieve templates information from S3.
+
+    Args:
+        url (str): The URL endpoint for retrieving templates info.
+
+    Returns:
+        response_data (Dict[str, Dict[str, Any]]): A dictionary \
+            containing dictionaries of templates information.
+    """
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get(url, timeout=10)
+        if response.status_code == 200:
+            response_data = response.json()
+            return response_data
+
+        response_data = response.json()
+        return response_data
